Add appending functionality to FileBuffer
FileBuffer can take a file object but the contents are only read when calling getContents(). This makes it difficult to append extra information on to the end of the buffer. It also means that the file reading will be called again if getContents is called a second time. Since this is part of the API of the object then this could easily happen. Since ultimately the contents will be written to a new file it doesn't matter when we do this read but it's better to do it just once.

To this end this patch changes the file read to the initialisation of the object meaning that we need only deal with strings internally. This makes it trivial to concatenate with another string. Note I have kept the external API method getContent in order to maintain backwards compatibility however it could easily be removed now.

	modified:   python/Ganga/GPIDev/Lib/File/FileBuffer.py